ffmpeg-vaapi: move common code to lib
