+++ lib/ffmpeg/vaapi/__init__.py
@@ -1,0 +1,5 @@
+###
+### Copyright (C) 2018-2020 Intel Corporation
+###
+### SPDX-License-Identifier: BSD-3-Clause
+###
