+++ test/ffmpeg-vaapi/decode/10bit/vp9.py
@@ -5,8 +5,8 @@
 ###
 
 from .....lib import *
-from ...util import *
-from ..decoder import DecoderTest
+from .....lib.ffmpeg.vaapi.util import *
+from .....lib.ffmpeg.vaapi.decoder import DecoderTest
 
 spec = load_test_spec("vp9", "decode", "10bit")
