+++ test/ffmpeg-vaapi/encode/vp8.py
@@ -5,8 +5,8 @@
 ###
 
 from ....lib import *
-from ..util import *
-from .encoder import EncoderTest
+from ....lib.ffmpeg.vaapi.util import *
+from ....lib.ffmpeg.vaapi.encoder import EncoderTest
 
 spec = load_test_spec("vp8", "encode")
