+++ test/ffmpeg-vaapi/vpp/contrast.py
@@ -5,8 +5,8 @@
 ###
 
 from ....lib import *
-from ..util import *
-from .vpp import VppTest
+from ....lib.ffmpeg.vaapi.util import *
+from ....lib.ffmpeg.vaapi.vpp import VppTest
 
 spec      = load_test_spec("vpp", "contrast")
 spec_r2r  = load_test_spec("vpp", "contrast", "r2r")
